+++ src/main/java/fr/alasdiablo/janoeo/blocks/BasicEndOre.java
@@ -1,16 +1,17 @@
-package fr.alasdiablo.janoeo.blocks.nether;
+package fr.alasdiablo.janoeo.blocks;
 
-import net.minecraft.block.RedstoneOreBlock;
+import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraftforge.common.ToolType;
 
-public class NetherRedstoneOre extends RedstoneOreBlock {
-    public NetherRedstoneOre(String registryName) {
+public class BasicEndOre extends Block {
+
+    public BasicEndOre(String registryName) {
         super(Properties.create(Material.ROCK)
                 .sound(SoundType.STONE)
-                .hardnessAndResistance(3f)
-                .harvestLevel(2)
+                .hardnessAndResistance(4f)
+                .harvestLevel(3)
                 .harvestTool(ToolType.PICKAXE)
         );
         this.setRegistryName(registryName);