+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndCoalOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.end;
-
-import fr.alasdiablo.janoeo.blocks.EndOre;
-
-public class EndCoalOre extends EndOre {
-    public EndCoalOre(String registryName) {
-        super(registryName);
-    }
-}