+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndDiamondOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.end;
-
-import fr.alasdiablo.janoeo.blocks.EndOre;
-
-public class EndDiamondOre extends EndOre {
-    public EndDiamondOre(String registryName) {
-        super(registryName);
-    }
-}