+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndEmeraldOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.end;
-
-import fr.alasdiablo.janoeo.blocks.EndOre;
-
-public class EndEmeraldOre extends EndOre {
-    public EndEmeraldOre(String registryName) {
-        super(registryName);
-    }
-}