+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndGoldOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.end;
-
-import fr.alasdiablo.janoeo.blocks.EndOre;
-
-public class EndGoldOre extends EndOre {
-    public EndGoldOre(String registryName) {
-        super(registryName);
-    }
-}