+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndLapisOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.end;
-
-import fr.alasdiablo.janoeo.blocks.EndOre;
-
-public class EndLapisOre extends EndOre {
-    public EndLapisOre(String registryName) {
-        super(registryName);
-    }
-}