+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndRedstoneOre.java
@@ -1,9 +1,18 @@
 package fr.alasdiablo.janoeo.blocks.end;
 
-import fr.alasdiablo.janoeo.blocks.EndOre;
+import net.minecraft.block.RedstoneOreBlock;
+import net.minecraft.block.SoundType;
+import net.minecraft.block.material.Material;
+import net.minecraftforge.common.ToolType;
 
-public class EndRedstoneOre extends EndOre {
+public class EndRedstoneOre extends RedstoneOreBlock {
     public EndRedstoneOre(String registryName) {
-        super(registryName);
+        super(Properties.create(Material.ROCK)
+                .sound(SoundType.STONE)
+                .hardnessAndResistance(4f)
+                .harvestLevel(2)
+                .harvestTool(ToolType.PICKAXE)
+        );
+        this.setRegistryName(registryName);
     }
 }