+++ src/main/java/fr/alasdiablo/janoeo/blocks/nether/NetherCoalOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.nether;
-
-import fr.alasdiablo.janoeo.blocks.NetherOre;
-
-public class NetherCoalOre extends NetherOre {
-    public NetherCoalOre(String registryName) {
-        super(registryName);
-    }
-}