+++ src/main/java/fr/alasdiablo/janoeo/blocks/nether/NetherDiamondOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.nether;
-
-import fr.alasdiablo.janoeo.blocks.NetherOre;
-
-public class NetherDiamondOre extends NetherOre {
-    public NetherDiamondOre(String registryName) {
-        super(registryName);
-    }
-}