+++ src/main/java/fr/alasdiablo/janoeo/blocks/nether/NetherEmeraldOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.nether;
-
-import fr.alasdiablo.janoeo.blocks.NetherOre;
-
-public class NetherEmeraldOre extends NetherOre {
-    public NetherEmeraldOre(String registryName) {
-        super(registryName);
-    }
-}