+++ src/main/java/fr/alasdiablo/janoeo/blocks/nether/NetherGoldOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.nether;
-
-import fr.alasdiablo.janoeo.blocks.NetherOre;
-
-public class NetherGoldOre extends NetherOre {
-    public NetherGoldOre(String registryName) {
-        super(registryName);
-    }
-}