+++ src/main/java/fr/alasdiablo/janoeo/blocks/nether/NetherIronOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.nether;
-
-import fr.alasdiablo.janoeo.blocks.NetherOre;
-
-public class NetherIronOre extends NetherOre {
-    public NetherIronOre(String registryName) {
-        super(registryName);
-    }
-}