+++ src/main/java/fr/alasdiablo/janoeo/blocks/nether/NetherLapisOre.java
@@ -1,9 +1,0 @@
-package fr.alasdiablo.janoeo.blocks.nether;
-
-import fr.alasdiablo.janoeo.blocks.NetherOre;
-
-public class NetherLapisOre extends NetherOre {
-    public NetherLapisOre(String registryName) {
-        super(registryName);
-    }
-}