+++ changelog.md
@@ -1,3 +1,8 @@
+### 1.2
++ fix missing xp drop
++ fix end/nether redstone ore
++ add tags for more compatiblity with other mods
++ rewrite/clean up the source code of the ore block
 ### 1.1
 + add end ore generation
 ### 1.0